+++ src/grandorgue/GOrgueSettings.cpp
@@ -132,6 +132,7 @@ GOrgueSettings::GOrgueSettings(wxString instance) :
 	PolyphonyLimit(this, wxT("General"), wxT("PolyphonyLimit"), 0, MAX_POLYPHONY, 2048),
 	Preset(this, wxT("General"), wxT("Preset"), 0, MAX_PRESET, 0),
 	ReleaseLength(this, wxT("General"), wxT("ReleaseLength"), 0, 3000, 0),
+	LanguageCode(this, wxT("General"), wxT("Language"), wxEmptyString),
 	BitsPerSample(this, wxT("General"), wxT("BitsPerSample"), 8, 24, 24),
 	Transpose(this, wxT("General"), wxT("Transpose"), -11, 11, 0),
 	MetronomeMeasure(this, wxT("Metronome"), wxT("Measure"), 0, 32, 4),
@@ -293,6 +294,26 @@ void GOrgueSettings::Load()
 		}
 		m_AudioDeviceConfig.push_back(conf);
 	}
+}
+
+int GOrgueSettings::GetLanguageId() const
+{
+  int langId = wxLANGUAGE_DEFAULT;
+  const wxString langCode = LanguageCode();
+
+  if (! langCode.IsEmpty())
+  {
+    const wxLanguageInfo* const langInfo =  wxLocale::FindLanguageInfo(langCode);
+
+    if (langInfo != NULL)
+      langId = langInfo->Language;
+  }
+  return langId;
+}
+
+void GOrgueSettings::SetLanguageId(int langId)
+{
+  LanguageCode(langId == wxLANGUAGE_DEFAULT ? wxString(wxEmptyString) : wxLocale::GetLanguageCanonicalName(langId));
 }
 
 unsigned GOrgueSettings::GetEventCount()