+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added capability of setting the UI language https://github.com/oleg68/GrandOrgue/issues/45
 - Fixed formatting of .md files
 # 0.3.1.2341-6.os (2021-07-07)
 - Changed the application icons. Now they are generated from a single svg file during the build https://github.com/oleg68/GrandOrgue/issues/20